fix: get correct value of total users

+++ src/main/java/io/github/saifalhaider/nahrain/nahrain_central_api/admin/getUsers/UsersDto.java
@@ -15,5 +15,5 @@ import java.util.List;
 public class UsersDto<T extends UserDto> {
     private List<T> users;
     private int totalPages;
-    private int totalNumberOfUsers;
+    private long totalNumberOfUsers;
 }
